+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_en.dart
@@ -85,6 +85,9 @@ class StreamChatLocalizationsEn extends GlobalStreamChatLocalizations {
   String get messageDeletedLabel => 'Message deleted';
 
   @override
+  String get systemMessageLabel => 'System Message';
+
+  @override
   String get editedMessageLabel => 'Edited';
 
   @override
@@ -629,4 +632,31 @@ class StreamChatLocalizationsEn extends GlobalStreamChatLocalizations {
   @override
   String get moderationReviewModalDescription =>
       '''Consider how your comment might make others feel and be sure to follow our Community Guidelines.''';
+
+  @override
+  String get emptyMessagePreviewText => '';
+
+  @override
+  String get voiceRecordingText => 'Voice Recording';
+
+  @override
+  String get audioAttachmentText => 'Audio';
+
+  @override
+  String get imageAttachmentText => 'Image';
+
+  @override
+  String get videoAttachmentText => 'Video';
+
+  @override
+  String get pollYouVotedText => 'You voted';
+
+  @override
+  String pollSomeoneVotedText(String username) => '$username voted';
+
+  @override
+  String get pollYouCreatedText => 'You created';
+
+  @override
+  String pollSomeoneCreatedText(String username) => '$username created';
 }