+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_ko.dart
@@ -84,6 +84,9 @@ class StreamChatLocalizationsKo extends GlobalStreamChatLocalizations {
   String get messageDeletedLabel => '메시지가 삭제되었습니다';
 
   @override
+  String get systemMessageLabel => '시스템 메시지';
+
+  @override
   String get editedMessageLabel => '편집됨';
 
   @override
@@ -609,5 +612,32 @@ class StreamChatLocalizationsKo extends GlobalStreamChatLocalizations {
 
   @override
   String get moderationReviewModalDescription =>
-      '귀하의 댓글이 다른 사람들에게 어떤 느낌을 줄 수 있는지 고려하고 커뮤니티 가이드라인을 준수하십시오.';
+      '''귀하의 댓글이 다른 사람들에게 어떤 영향을 미칠 수 있는지 고려하고 커뮤니티 가이드라인을 준수하세요.''';
+
+  @override
+  String get emptyMessagePreviewText => '';
+
+  @override
+  String get voiceRecordingText => '음성 녹음';
+
+  @override
+  String get audioAttachmentText => '오디오';
+
+  @override
+  String get imageAttachmentText => '이미지';
+
+  @override
+  String get videoAttachmentText => '비디오';
+
+  @override
+  String get pollYouVotedText => '투표했습니다';
+
+  @override
+  String pollSomeoneVotedText(String username) => '$username님이 투표했습니다';
+
+  @override
+  String get pollYouCreatedText => '생성했습니다';
+
+  @override
+  String pollSomeoneCreatedText(String username) => '$username님이 생성했습니다';
 }