+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_no.dart
@@ -85,6 +85,9 @@ class StreamChatLocalizationsNo extends GlobalStreamChatLocalizations {
   String get messageDeletedLabel => 'Melding slettet';
 
   @override
+  String get systemMessageLabel => 'Systemmelding';
+
+  @override
   String get editedMessageLabel => 'Redigert';
 
   @override
@@ -619,5 +622,32 @@ class StreamChatLocalizationsNo extends GlobalStreamChatLocalizations {
 
   @override
   String get moderationReviewModalDescription =>
-      '''Tenk på hvordan kommentaren din kan få andre til å føle seg, og sørg for å følge retningslinjene for fellesskapet vårt.''';
+      '''Tenk på hvordan kommentaren din kan få andre til å føle seg og sørg for å følge retningslinjene for fellesskapet.''';
+
+  @override
+  String get emptyMessagePreviewText => '';
+
+  @override
+  String get voiceRecordingText => 'Taleopptak';
+
+  @override
+  String get audioAttachmentText => 'Lyd';
+
+  @override
+  String get imageAttachmentText => 'Bilde';
+
+  @override
+  String get videoAttachmentText => 'Video';
+
+  @override
+  String get pollYouVotedText => 'Du stemte';
+
+  @override
+  String pollSomeoneVotedText(String username) => '$username stemte';
+
+  @override
+  String get pollYouCreatedText => 'Du opprettet';
+
+  @override
+  String pollSomeoneCreatedText(String username) => '$username opprettet';
 }